requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
sqlalchemy==2.0.41
asyncpg==0.30.0
